requirements.txt
pandas>=2.3.3
numpy>=2.3.4
scikit-learn>=1.7.2
sentence-transformers>=5.1.2
faiss-cpu>=1.12.0
python-dotenv>=1.2.1
tqdm>=4.67.1

# Imbalanced learning library for handling class imbalance
imbalanced-learn>=0.14.0

# LLM + LangChain stack (ACTIVE) - Latest versions as of Oct 2025
langchain>=1.0.3
langchain-core>=1.0.2
langchain-openai>=1.0.1

# show-off, not used in main
openai>=2.6.1

pydantic>=2.12.3

# Retry library for robust API calls
tenacity>=9.0.0
